调整架构

+++ src/bot_config.py
@@ -55,7 +55,7 @@ class RuntimeConfig:
     """机器人运行时的控制参数。"""
 
     max_try_combat_count: int = 3  # 单次战斗尝试的最大次数
-    max_ops_per_second: int = 2  # 每秒允许的最大操作频率（防止动作过快被检测）
+    max_ops_per_second: float = 2  # 每秒允许的最大操作频率（防止动作过快被检测）
 
 
 @dataclass(frozen=True)
